# control-api/docs/swagger.yaml
basePath: /
definitions:
...
  gateway.channelDTO:
    properties:
      amount:
        type: string
      asset_id:
        type: string
      asset_symbol:
        type: string
      caip2_chain_id:
        type: string
      channel_config:
        additionalProperties:
          type: string
        type: object
      code:
        type: string
      contract_address:
        type: string
      decimals:
        type: integer
      enabled:
        type: boolean
      facilitator_url:
        type: string
      method:
        type: string
      payment_method_id:
        type: string
      payout_address:
        type: string
      protocol:
        type: string
      scheme:
        type: string
    type: object
  gateway.errorResponse:
    properties:
      error:
        type: string
    type: object
  gateway.resolveRouteResponse:
    properties:
      allow_unmatched:
        type: boolean
      auth_header_name:
        type: string
      auth_header_value:
        type: string
      bazaar:
        type: object
      description:
        type: string
      free:
        type: boolean
      inbound_path:
        type: string
      mime_type:
        type: string
      name:
        type: string
      outbound_route_id:
        type: string
      payment_channels:
        items:
          $ref: '#/definitions/gateway.channelDTO'
        type: array
      price:
        type: string
      project_id:
        type: string
      target:
        type: string
    type: object
...
host: localhost:9090
info:
  contact: {}
  description: Control plane for xpaywall — manages projects, routes, users, payment
    channels, and request logs.
  title: xpaywall Control API
  version: "1.0"
paths:
...
  /proxy/resolve/{path}:
    get:
      parameters:
      - description: Path in format {username}/{projectSlug}/{inboundPath}
        in: path
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gateway.resolveRouteResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gateway.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gateway.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Resolve route
      tags:
      - gateway
securityDefinitions:
  ApiKeyAuth:
    description: Internal API key shared between xgateway and control-api
    in: header
    name: X-Api-Key
    type: apiKey
  BearerAuth:
    description: 'JWT token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"